requirements.txt
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
h11==0.14.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.9.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
preshed==3.0.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
rich==13.7.1
selenium==4.23.1
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thefuzz==0.22.1
thinc==8.2.5
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
